+++ CoreUtilities/LJCDataUtility/LJCDataUtilityDAL/Manager/DataKeymanager.cs
@@ -134,7 +134,7 @@ namespace LJCDataUtilityDAL
 
     // Retrieves a record with the supplied value.
     /// <include path='items/RetrieveWithID/*' file='../../LJCDocLib/Common/Manager.xml'/>
-    public DataKey RetrieveWithID(int id, List<string> propertyNames = null)
+    public DataKey RetrieveWithID(long id, List<string> propertyNames = null)
     {
       DataKey retValue;
 
@@ -148,7 +148,7 @@ namespace LJCDataUtilityDAL
 
     // Retrieves a record with the supplied unique values.
     /// <include path='items/RetrieveWithName/*' file='../../LJCDocLib/Common/Manager.xml'/>
-    public DataKey RetrieveWithUnique(int dataTableID, string name
+    public DataKey RetrieveWithUnique(long dataTableID, string name
       , List<string> propertyNames = null)
     {
       DataKey retValue;
@@ -166,7 +166,7 @@ namespace LJCDataUtilityDAL
 
     // Gets the ID key columns.
     /// <include path='items/GetIDKey/*' file='../../LJCDocLib/Common/Manager.xml'/>
-    public DbColumns IDKey(int id)
+    public DbColumns IDKey(long id)
     {
       // Add(columnName, propertyName = null, renameAs = null
       //   , datatypeName = "String", caption = null);
@@ -180,7 +180,7 @@ namespace LJCDataUtilityDAL
 
     // Gets the ID key columns.
     /// <include path='items/GetIDKey/*' file='../../LJCDocLib/Common/Manager.xml'/>
-    public DbColumns ParentKey(int parentID)
+    public DbColumns ParentKey(long parentID)
     {
       var retValue = new DbColumns()
       {
@@ -191,7 +191,7 @@ namespace LJCDataUtilityDAL
 
     // Gets the ID key columns.
     /// <include path='items/GetNameKey/*' file='../../LJCDocLib/Common/Manager.xml'/>
-    public DbColumns UniqueKey(int dataTableID, string name)
+    public DbColumns UniqueKey(long dataTableID, string name)
     {
       // Needs cast for string to select the correct Add overload.
       var retValue = new DbColumns()
